# lib/api-zod/src/generated/types/achievementResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PawPlay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AchievementResponseMetadata } from "./achievementResponseMetadata";

export interface AchievementResponse {
  id: string;
  dogId: string;
  /** @nullable */
  userId?: string | null;
  type: string;
  metadata?: AchievementResponseMetadata;
  earnedAt: string;
}